changed: reducers structure
