+++ src/pages/private/user/index.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function User(props) {
+    return (
+        <div>
+            <h1>User</h1>
+        </div>
+    );
+}
