+++ src/pages/public/sign-in/saga.js
@@ -1,9 +1,7 @@
 import {put, takeEvery, call} from 'redux-saga/effects';
 import {ASYNC_SET_TOKEN, ASYNC_SET_USER, asyncSetUser, setUser} from "./reducer";
 import {getDataFromApi, getUserData} from '../../../utils/API';
-import {setToken} from "./reducer";
-
-
+import {setToken} from "../../reducer";
 
 function* setTokenWorker(action) {
     const {data} = yield call(getDataFromApi, action.payload);
@@ -11,8 +9,6 @@ function* setTokenWorker(action) {
     yield put(setToken(data.accessToken));
     yield put(asyncSetUser(data.accessToken));
 }
-
-
 
 function* setUserWorker(action) {
     const {data} = yield call(getUserData, action.payload);
