+++ src/pages/reducer.js
@@ -1,3 +1,29 @@
 import publicReducers from './public/reducer';
+import {combineReducers} from "redux";
 
-export default publicReducers;
+export const SET_TOKEN = "SET_TOKEN";
+
+const tokenInitialState = {
+    accessToken: localStorage.getItem('token') || ''
+};
+
+function tokenReducer(state = tokenInitialState, action) {
+    switch (action.type) {
+        case SET_TOKEN:
+            return {token: action.payload};
+        default:
+            return state;
+    }
+}
+
+export function setToken(token) {
+    return {
+        type: SET_TOKEN,
+        payload: token
+    };
+}
+
+export default combineReducers({
+    public: publicReducers,
+    token: tokenReducer
+});
